separate benches from tests
